refactor: SSR의 cookie 때문에 api 함수의 인자로 undefined 넣었던 부분을 optional로 수정

+++ src/features/auth/hooks/useAuth.ts
@@ -78,7 +78,14 @@ export const useAuth = () => {
   };
 
   const logout = () => {
-    logoutMutation.mutate(localStorage.getItem('accessToken'));
+    const accessToken = localStorage.getItem('accessToken');
+    if (!accessToken) {
+      // eslint-disable-next-line no-console
+      console.error('로그아웃 에러');
+      alert('로그아웃을 다시 시도해주십시요.');
+      return;
+    }
+    logoutMutation.mutate(accessToken);
     localStorage.removeItem('accessToken');
     localStorage.removeItem('user');
     dispatch(removeUser());
